Add fps meter to preference
