+++ src/index.tsx
@@ -54,25 +54,3 @@ async function fetchConfig() {
 fetchConfig().then(() => {
     console.log("Configuration complete. Rendering frontend");
 });
-
-// Add FPS meter for development
-if (process.env.NODE_ENV === "development") {
-    import("stats-js")
-        .then(({default: Stats}) => {
-            const stats = new Stats();
-            stats.showPanel(0); // 0: fps, 1: ms, 2: mb, 3+: custom
-            document.body.appendChild(stats.dom);
-            function animate() {
-                stats.begin();
-                // monitored code goes here
-                stats.end();
-                requestAnimationFrame(animate);
-            }
-            requestAnimationFrame(animate);
-            stats.dom.style.right = "0";
-            stats.dom.style.left = "initial";
-        })
-        .catch(err => {
-            console.log(err);
-        });
-}
