+++ src/models/preferences_schema_2.json
@@ -293,6 +293,15 @@
             "type": "number",
             "minimum": 1,
             "default": 2
+        },
+        "statsPanelEnabled": {
+            "type": "boolean",
+            "default": false
+        },
+        "statsPanelMode": {
+            "type": "integer",
+            "minimum": 0,
+            "default": 0
         }
     }
 }
